requirements.txt
llama-index
llama-index-llms-gemini
llama-index-llms-ollama
llama-index-embeddings-gemini
llama-index-embeddings-huggingface
llama-index-vector-stores-qdrant
llama-index-postprocessor-rerank-bge-rerank
qdrant-client
streamlit
python-dotenv
pdf2image
pdfplumber
Pillow
trafilatura
beautifulsoup4
sentence-transformers
ollama
google-generativeai
